#11 Move pulling batch mapping in separate file
