+++ debatch_manager/src/debatcher.py
@@ -8,8 +8,6 @@ __email__ = "m.gafarova@eora.ru"
 from typing import List, Iterator, Optional, Iterable
 from itertools import repeat
 
-import plyvel  # type: ignore
-from loguru import logger
 
 from shared_modules.utils import uuid4_string_generator
 import src.data_models as dm
@@ -51,34 +49,3 @@ def debatch(
             response_objects.append(new_response_object)
 
     return response_objects
-
-
-def pull_batch_mapping(config: dm.Config, batch: dm.ResponseBatch) -> dm.BatchMapping:
-    """
-    Retrieve dm.BatchMapping from database
-
-    Parameters
-    ---------
-    config:
-        Config of database
-    batch:
-        Response Batch object
-    """
-    # Open conndection
-    try:
-        database = plyvel.DB(
-            config.db_file,
-        )
-    except plyvel._plyvel.IOError as exc:
-        raise IOError() from exc
-    # Pull batch mapping from database
-    batch_mapping_bytes = database.get(bytes(batch.uid, "utf-8"))
-    batch_mapping = dm.BatchMapping.from_key_value(
-        (bytes(batch.uid, "utf-8"), batch_mapping_bytes)
-    )
-
-    # Delete mapping
-    database.delete(bytes(batch.uid, "utf-8"))
-    database.close()
-
-    return batch_mapping
